ADDED : Refresh feature CHANGES : Email section into component
