+++ client/client_main/src/index.css
@@ -67,16 +67,3 @@ body {
 .app_section_options {
   flex: 2;
 }
-
-.app_section_emails {
-  flex: 8;
-  padding: 3em 5em;
-  overflow-y: auto;
-}
-
-.app_section_emails_title {
-  font-family: "Space Grotesk", sans-serif;
-  font-weight: 600;
-  letter-spacing: 2px;
-  font-size: 24px;
-}
